# core: optimize tap-targets audit