--- a/lighthouse-core/audits/seo/tap-targets.js
+++ b/lighthouse-core/audits/seo/tap-targets.js
@@ -36,7 +36,7 @@ const UIStrings = {
   /** Label of a table column that identifies a tap target (like a link or button) that overlaps with another tap target. */
   overlappingTargetHeader: 'Overlapping Target',
   /** Explanatory message stating that there was a failure in an audit caused by the viewport meta tag not being optimized for mobile screens, which caused tap targets like buttons and links to be too small to tap on. */
-  /* eslint-disable max-len */
+  /* eslint-disable-next-line max-len */
   explanationViewportMetaNotOptimized: 'Tap targets are too small because there\'s no viewport meta tag optimized for mobile screens',
   /** Explanatory message stating that a certain percentage of the tap targets (like buttons and links) on the page are of an appropriately large size. */
   displayValue: '{decimalProportion, number, percent} appropriately sized tap targets',
@@ -59,7 +59,7 @@ function getBoundedTapTargets(targets) {
   return targets.map(tapTarget => {
     return {
       tapTarget,
-      boundingRect: getBoundingRectWithPadding(tapTarget.clientRects, FINGER_SIZE_PX),
+      paddedBoundsRect: getBoundingRectWithPadding(tapTarget.clientRects, FINGER_SIZE_PX),
     };
   });
 }
@@ -101,7 +101,7 @@ function getAllOverlapFailures(tooSmallTargets, allTargets) {
         continue;
       }
 
-      if (!rectsTouchOrOverlap(target.boundingRect, maybeOverlappingTarget.boundingRect)) {
+      if (!rectsTouchOrOverlap(target.paddedBoundsRect, maybeOverlappingTarget.paddedBoundsRect)) {
         // Bounding boxes (padded with half FINGER_SIZE_PX) don't overlap, skip.
         continue;
       }
@@ -334,7 +334,7 @@ module.exports.UIStrings = UIStrings;
 }} TapTargetOverlapFailure */
 
 /** @typedef {{
-  boundingRect: LH.Artifacts.Rect;
+  paddedBoundsRect: LH.Artifacts.Rect;
   tapTarget: LH.Artifacts.TapTarget;
 }} BoundedTapTarget */