--- a/lighthouse-core/test/lib/rect-helpers-test.js
+++ b/lighthouse-core/test/lib/rect-helpers-test.js
@@ -13,6 +13,7 @@ const {
   getLargestRect,
   getRectAtCenter,
   allRectsContainedWithinEachOther,
+  getBoundingRectWithPadding,
 } = require('../../lib/rect-helpers');
 
 describe('Rect Helpers', () => {
@@ -82,6 +83,37 @@ describe('Rect Helpers', () => {
       const rect1 = addRectTopAndBottom({x: 0, y: 0, width: 100, height: 100});
       const rect2 = addRectTopAndBottom({x: 200, y: 200, width: 20, height: 20});
       expect(allRectsContainedWithinEachOther([rect1], [rect2])).toBe(false);
+    });
+  });
+
+  describe('#getBoundingRectWithPadding', () => {
+    it('throws an error if no rects are passed in', () => {
+      expect(() => getBoundingRectWithPadding([])).toThrow('No rects');
+    });
+
+    it('pads rect with half minimum size on all sides', () => {
+      const rect = addRectTopAndBottom({x: 0, y: 0, width: 0, height: 0});
+      const minimumSize = 20;
+      const expectedPaddedBounds = addRectTopAndBottom({x: -10, y: -10, width: 20, height: 20});
+
+      expect(getBoundingRectWithPadding([rect], minimumSize)).toEqual(expectedPaddedBounds);
+    });
+
+    it('pads the bounding box of two rects with half of the minimum size', () => {
+      const rect1 = addRectTopAndBottom({x: 0, y: 0, width: 10, height: 10});
+      const rect2 = addRectTopAndBottom({x: 50, y: 50, width: 10, height: 10}); // in the middle somewhere
+      const rect3 = addRectTopAndBottom({x: 100, y: 100, width: 10, height: 10});
+
+      const minimumSize = 20;
+      const expectedPaddedBounds = addRectTopAndBottom({
+        x: -minimumSize / 2,
+        y: -minimumSize / 2,
+        width: 110 + minimumSize,
+        height: 110 + minimumSize,
+      });
+
+      const rects = [rect1, rect2, rect3];
+      expect(getBoundingRectWithPadding(rects, minimumSize)).toEqual(expectedPaddedBounds);
     });
   });
 });