@mattzeunert 2019-02-01
Nice, that's a pretty significant speed-up 😍

> allRectsContainedWithinEachOther is about 10% of the runtime of audit of the wikipedia page and doesn't appear to ever do anything. Conversation in #5846 (comment) made it seem like it might only be needed when we add position:absolute?

Yeah, right now we just ignore `position: absolute` tap targets, and the gatherer already ignores elements that have an ancestor that's a tap target. We could only run that check if either tap target is actually `position: absolute`.